chore: refactor my nodes
