+++ src/app/dashboard/pages/nodes/components/my-nodes-create-modal/my-nodes-create-modal.component.scss
@@ -6,5 +6,5 @@
 
 .footer {
   display: flex;
-  justify-content: space-between;
+  justify-content: flex-end;
 }
