+++ src/app/dashboard/pages/nodes/components/node-details/node-details.component.html
@@ -54,7 +54,6 @@
         <nz-space-item>
           <button
             nz-button
-            nzSize="small"
             nzType="primary"
             (click)="onUpdateMyNodesClick($event)"
           >
@@ -64,7 +63,6 @@
         <nz-space-item>
           <button
             nz-button
-            nzSize="small"
             nzType="primary"
             [disabled]="isAddedToMyNodes$ | async"
             (click)="onAddToMyNodesClick($event)"
@@ -82,7 +80,6 @@
       <button
         class="node-manager-btn"
         nz-button
-        nzSize="small"
         nzType="primary"
         [nzLoading]="isLoadingNodeManager$ | async"
         (click)="onNodeManagerClick($event)"
@@ -95,4 +92,25 @@
 
 <ng-template #loading>
   <nz-spin nzSimple></nz-spin>
+</ng-template>
+
+<ng-template #nodeDetailsModalTitleTpl>
+  <app-modal-header>
+    <nz-space>
+      <nz-space-item>
+        <h4 nz-typography nzType="secondary">
+          Select node
+        </h4>
+      </nz-space-item>
+      <nz-space-item>
+        <app-popover-helper
+          placement="right"
+          [popoverContent]="nodeDetailsHelperTpl"
+        ></app-popover-helper>
+      </nz-space-item>
+    </nz-space>
+  </app-modal-header>
+  <ng-template #nodeDetailsHelperTpl>
+    Select node to manage it
+  </ng-template>
 </ng-template>
